WIP. Refactor some util functions into ResourceUtils, add the preference pane code to the OSXApplicationBuilder.
