+++ src/com/inet/gradle/setup/dmg/DmgBuilder.java
@@ -30,9 +30,6 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.attribute.FileAttribute;
 import java.util.ArrayList;
-import java.util.Enumeration;
-import java.util.zip.ZipEntry;
-import java.util.zip.ZipFile;
 
 import javax.imageio.ImageIO;
 
@@ -43,6 +40,7 @@ import com.inet.gradle.setup.AbstractBuilder;
 import com.inet.gradle.setup.DesktopStarter;
 import com.inet.gradle.setup.Service;
 import com.inet.gradle.setup.SetupBuilder;
+import com.inet.gradle.setup.util.ResourceUtils;
 import com.inet.gradle.setup.util.XmlFileBuilder;
 
 /**
@@ -75,7 +73,8 @@ public class DmgBuilder extends AbstractBuilder<Dmg> {
 
         try {
         	tmp = Files.createTempDirectory("SetupBuilder", new FileAttribute[0]);
-        	OSXApplicationBuilder applicationBuilder = new OSXApplicationBuilder( task, setup, fileResolver );
+        	System.out.println( "created temporary directory: " + tmp);
+        	OSXApplicationBuilder applicationBuilder = new OSXApplicationBuilder( task, setup, fileResolver, tmp );
         	
         	// Build all services 
         	for (Service service : setup.getServices() ) {
@@ -90,7 +89,7 @@ public class DmgBuilder extends AbstractBuilder<Dmg> {
             title = setup.getSetupName();
             applicationName = setup.getBaseName();
             // applicationName = setup.getApplication();
-            imageSourceRoot = buildDir.toString() + "/" + applicationName + ".app";
+            imageSourceRoot = buildDir.toString() + "/" + setup.getApplication() + ".app";
             iconFile = applicationBuilder.getApplicationIcon();
 
         	if ( !setup.getServices().isEmpty() ) {
@@ -154,7 +153,7 @@ public class DmgBuilder extends AbstractBuilder<Dmg> {
         	String launchdScriptName = setup.getMainClass();
             InputStream input = getClass().getResourceAsStream( resource );
             ByteArrayOutputStream bos = new ByteArrayOutputStream();
-            copyData(input, bos);
+            ResourceUtils.copyData(input, bos);
             input.close();
 
             String script = new String( bos.toByteArray(), StandardCharsets.UTF_8 );
@@ -185,92 +184,6 @@ public class DmgBuilder extends AbstractBuilder<Dmg> {
     	}
     }
     
-    /**
-     * Copy the data from the input to the output
-     * @param input the input data
-     * @param output the target
-     * @throws IOException if IO issues
-     */
-    public static void copyData( InputStream input, OutputStream output ) throws IOException{
-        byte[] buf = new byte[4096];
-        int len;
-        while ((len = input.read(buf)) > 0) {
-            output.write(buf, 0, len);
-        }
-    }
-	
-	/**
-	 * * Unpack the preference pane from the SetupBuilder,
-	 * * Modify the icon and text
-	 * * Put into the main application bundle
-	 * 
-	 * @throws Throwable
-	 */
-	private void createPreferencePane() throws Throwable {
-
-		// Unpack
-		File setPrefpane = new File(tmp.toFile(), "packages/SetupBuilderOSXPrefPane.prefPane.zip");
-
-		InputStream input = getClass().getResourceAsStream("service/SetupBuilderOSXPrefPane.prefPane.zip");
-		FileOutputStream output = new FileOutputStream( setPrefpane );
-        copyData(input, output);
-        input.close();
-        output.close();
-
-        File resourcesOutput = new File(buildDir, applicationName + ".app/Contents/Resources");
-		unZipIt(setPrefpane, resourcesOutput);
-
-		// Rename to app-name
-        File prefPaneLocation = new File(resourcesOutput, applicationName + ".prefPane");
-
-        // rename helper Tool
-        File prefPaneContents = new File(resourcesOutput, "SetupBuilderOSXPrefPane.prefPane/Contents");
-		Files.copy(iconFile.toPath(), new File(prefPaneContents, "Resources/SetupBuilderOSXPrefPane.app/Contents/Resources/applet.icns").toPath(), java.nio.file.StandardCopyOption.REPLACE_EXISTING);
-		Files.move(new File(prefPaneContents, "MacOS/SetupBuilderOSXPrefPane").toPath(), new File( prefPaneContents, "MacOS/"+ applicationName ).toPath(),  java.nio.file.StandardCopyOption.REPLACE_EXISTING );
-		Files.move(new File(prefPaneContents, "Resources/SetupBuilderOSXPrefPane.app").toPath(), new File(prefPaneContents, "Resources/"+ applicationName +".app").toPath(),  java.nio.file.StandardCopyOption.REPLACE_EXISTING );
-
-		// Make executable
-        ArrayList<String> command = new ArrayList<>();
-        command.add( "chmod" );
-        command.add( "a+x" );
-        command.add(  new File(prefPaneContents, "Resources/"+ applicationName +".app/Contents/MacOS/applet").getAbsolutePath() );
-    	exec( command );
-    	
-		// Rename prefPane
-		Files.move(new File(resourcesOutput, "SetupBuilderOSXPrefPane.prefPane").toPath(), prefPaneLocation.toPath(),  java.nio.file.StandardCopyOption.REPLACE_EXISTING );
-        Files.delete(setPrefpane.toPath());
-
-		// Copy Icon
-		Files.copy(iconFile.toPath(), new File(prefPaneLocation, "Contents/Resources/ProductIcon.icns").toPath(), java.nio.file.StandardCopyOption.REPLACE_EXISTING);
-		
-		// Patch Info.plist
-		File prefPanePLIST = new File(prefPaneLocation, "Contents/Info.plist");
-		setPlistProperty( prefPanePLIST, ":CFBundleIdentifier", setup.getMainClass() + ".prefPane" );
-		setPlistProperty( prefPanePLIST, ":CFBundleName", title + " Preference Pane" );
-		setPlistProperty( prefPanePLIST, ":CFBundleExecutable", applicationName );
-		setPlistProperty( prefPanePLIST, ":NSPrefPaneIconLabel", title );
-
-		setPlistProperty( prefPanePLIST, ":CFBundleExecutable", applicationName );
-		
-		File servicePLIST = new File(prefPaneLocation, "Contents/Resources/service.plist");
-		setPlistProperty( servicePLIST, ":Name", title );
-		setPlistProperty( servicePLIST, ":Label", setup.getMainClass() );
-		setPlistProperty( servicePLIST, ":Program", "/Library/" + applicationName + "/" + applicationName + ".app/Contents/MacOS/" + setup.getBaseName() );
-		setPlistProperty( servicePLIST, ":Description", setup.getServices().get(0).getDescription() );
-		setPlistProperty( servicePLIST, ":Version", setup.getVersion() );
-	}
-	
-	private void setPlistProperty(File plist, String property, String value) {
-		
-		// Set Property in plist file
-		// /usr/libexec/PlistBuddy -c "Set PreferenceSpecifiers:19:Titles:0 $buildDate" "$BUNDLE/Root.plist"
-        ArrayList<String> command = new ArrayList<>();
-        command.add( "/usr/libexec/PlistBuddy" );
-        command.add( "-c" );
-        command.add( "Set " + property +  " " + value );
-        command.add( plist.getAbsolutePath() );
-    	exec( command );
-	}
 
     private void createPackageFromApp() throws Throwable {
 
@@ -282,7 +195,6 @@ public class DmgBuilder extends AbstractBuilder<Dmg> {
 		patchServiceFiles( "scripts/preinstall", new File(tmp.toFile(), "scripts/preinstall") );
 		patchServiceFiles( "scripts/postinstall", new File(tmp.toFile(), "scripts/postinstall") );
 
-		createPreferencePane();
         extractApplicationInformation();
     	createAndPatchDistributionXML();
 
@@ -453,7 +365,7 @@ public class DmgBuilder extends AbstractBuilder<Dmg> {
     private void applescript() throws IOException {
         InputStream input = getClass().getResourceAsStream( "applescript.txt" );
         ByteArrayOutputStream bos = new ByteArrayOutputStream();
-        copyData(input, bos);
+        ResourceUtils.copyData(input, bos);
         input.close();
 
         String script = new String( bos.toByteArray(), StandardCharsets.UTF_8 );
@@ -491,45 +403,4 @@ public class DmgBuilder extends AbstractBuilder<Dmg> {
         exec( command );
     }
 
-	/**
-	 * Unzip it
-	 * 
-	 * @param file input zip file
-	 * @param output zip file output folder
-	 */
-	private void unZipIt(File file, File folder) {
-
-		try {
-
-			// create output directory is not exists
-			if (!folder.exists()) {
-				folder.mkdir();
-			}
-
-			// get the zip file content
-			ZipFile zipFile = new ZipFile(file);
-			Enumeration<? extends ZipEntry> entries = zipFile.entries();
-			while (entries.hasMoreElements()) {
-				ZipEntry zipEntry = (ZipEntry) entries.nextElement();
-				
-				String fileName = zipEntry.getName();
-				if (zipEntry.isDirectory() ) {
-					new File( folder, fileName).mkdir();
-				} else {
-					
-					File target = new File(folder, fileName);
-					InputStream inputStream = zipFile.getInputStream(zipEntry);
-					FileOutputStream output = new FileOutputStream(target);
-					copyData(inputStream, output);
-					output.close();
-					inputStream.close();
-				}
-			}
-			
-			zipFile.close();
-
-		} catch (IOException ex) {
-			ex.printStackTrace();
-		}
-	}
 }
