+++ src/com/inet/gradle/setup/util/ResourceUtils.java
@@ -16,10 +16,15 @@
 package com.inet.gradle.setup.util;
 
 import java.io.File;
+import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.OutputStream;
 import java.nio.file.Files;
 import java.nio.file.StandardCopyOption;
+import java.util.Enumeration;
+import java.util.zip.ZipEntry;
+import java.util.zip.ZipFile;
 
 /**
  * Some Utils for working with resources.
@@ -42,5 +47,61 @@ public class ResourceUtils {
             Files.copy( input, file.toPath(), StandardCopyOption.REPLACE_EXISTING );
         }
         return file;
+    }
+
+	/**
+	 * Unzip it
+	 * 
+	 * @param file input zip file
+	 * @param output zip file output folder
+	 */
+	public static void unZipIt(File file, File folder) {
+
+		try {
+
+			// create output directory is not exists
+			if (!folder.exists()) {
+				folder.mkdir();
+			}
+
+			// get the zip file content
+			ZipFile zipFile = new ZipFile(file);
+			Enumeration<? extends ZipEntry> entries = zipFile.entries();
+			while (entries.hasMoreElements()) {
+				ZipEntry zipEntry = (ZipEntry) entries.nextElement();
+				
+				String fileName = zipEntry.getName();
+				if (zipEntry.isDirectory() ) {
+					new File( folder, fileName).mkdir();
+				} else {
+					
+					File target = new File(folder, fileName);
+					InputStream inputStream = zipFile.getInputStream(zipEntry);
+					FileOutputStream output = new FileOutputStream(target);
+					copyData(inputStream, output);
+					output.close();
+					inputStream.close();
+				}
+			}
+			
+			zipFile.close();
+
+		} catch (IOException ex) {
+			ex.printStackTrace();
+		}
+	}
+	
+	/**
+     * Copy the data from the input to the output
+     * @param input the input data
+     * @param output the target
+     * @throws IOException if IO issues
+     */
+    public static void copyData( InputStream input, OutputStream output ) throws IOException{
+        byte[] buf = new byte[4096];
+        int len;
+        while ((len = input.read(buf)) > 0) {
+            output.write(buf, 0, len);
+        }
     }
 }
